fix hw3
