+++ controllers/contacts/deleteById.js
@@ -6,7 +6,7 @@ const deleteById = async (req, res) => {
   const { id } = req.params;
   const result = await Contact.findByIdAndRemove(id);
   if (!result) {
-    throw ServiceError(404);
+    throw new ServiceError(404);
   }
   res.json({
     message: "contact deleted",
