+++ controllers/contacts/getById.js
@@ -6,7 +6,7 @@ const getById = async (req, res) => {
   const { id } = req.params;
   const result = await Contact.findById(id, "-createdAt -updatedAt");
   if (!result) {
-    throw ServiceError(404);
+    throw new ServiceError(404);
   }
   res.json(result);
 };
