+++ controllers/contacts/updateById.js
@@ -6,7 +6,7 @@ const updateById = async (req, res) => {
   const { id } = req.params;
   const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
   if (!result) {
-    throw ServiceError(404);
+    throw new ServiceError(404);
   }
   res.json(result);
 };
