+++ controllers/contacts/updateFavorite.js
@@ -6,7 +6,7 @@ const updateFavorite = async (req, res) => {
   const { id } = req.params;
   const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
   if (!result) {
-    throw ServiceError(404);
+    throw new ServiceError(404);
   }
   res.json(result);
 };
