+++ middlewares/index.js
@@ -1,7 +1,9 @@
-const validation = require("./validation");
-const validationId = require("./validationId");
+const validation = require('./validation');
+const validationId = require('./validationId');
+const validationJoi = require('./validationJoi')
 
 module.exports = {
   validation,
   validationId,
+  validationJoi
 };
