+++ routes/api/contacts.js
@@ -1,30 +1,30 @@
 const express = require('express');
 const router = express.Router();
-const ctrl = require("../../controllers/contacts");
+const {getAll, getById, add, updateById, updateFavorite, deleteById} = require("../../controllers/contacts");
 const { controllerWrapper } = require("../../helpers");
 const { validation, validationId } = require("../../middlewares");
 const { schemas } = require("../../models/contacts");
 
-router.get("/", controllerWrapper(ctrl.getAll));
+router.get("/", controllerWrapper(getAll));
 
-router.get("/:id", validationId, controllerWrapper(ctrl.getById));
+router.get("/:id", validationId, controllerWrapper(getById));
 
-router.post("/", controllerWrapper(ctrl.add));
+router.post("/", controllerWrapper(add));
 
 router.put(
   "/:id",
   validationId,
   validation(schemas.add),
-  controllerWrapper(ctrl.updateById)
+  controllerWrapper(updateById)
 );
 
 router.patch(
   "/:id/favorite",
   validationId,
   validation(schemas.updateFavorite),
-  controllerWrapper(ctrl.updateFavorite)
+  controllerWrapper(updateFavorite)
 );
 
-router.delete("/:id", validationId, controllerWrapper(ctrl.deleteById));
+router.delete("/:id", validationId, controllerWrapper(deleteById));
 
 module.exports = router
